requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
tensorflow>=2.8.0
torch>=1.11.0
pyomo>=6.4.0
scipy>=1.7.0
requests>=2.26.0
pyyaml>=6.0
pytest>=6.2.0
tqdm>=4.62.0
jupyter>=1.0.0